Network.Wai.Middleware.AcceptOverride

Documentation

This Middleware provides a way for the request itself to tell the server to override the "Accept" header by looking for the "_accept" query parameter in the query string and inserting or replacing the "Accept" header with that string.

For example:

?_accept=SomeValue

This will result in "Accept: SomeValue" being set in the request as a header, and all other previous "Accept" headers will be removed from the request.

acceptOverride :: Middleware