Network.Wai.Middleware.RequestLogger

Basic stdout logging

logStdout :: Middleware

Production request logger middleware. Implemented on top of logCallback, but prints to stdout

logStdoutDev :: Middleware

Development request logger middleware. Implemented on top of logCallbackDev, but prints to stdout

Flushes stdout on each request, which would be inefficient in production use. Use logStdout in production.

Create more versions

mkRequestLogger :: RequestLoggerSettings -> IO Middleware

data RequestLoggerSettings

outputFormat :: RequestLoggerSettings -> OutputFormat

Default value: Detailed True.

autoFlush :: RequestLoggerSettings -> Bool

Only applies when using the Handle constructor for destination.

Default value: True.

destination :: RequestLoggerSettings -> Destination

Default: Handle stdout.

data OutputFormat

Constructors

Apache IPAddrSource

Detailed Bool
use colors?

data Destination

Constructors

Handle Handle

Logger Logger

Callback Callback

type Callback = [LogStr] -> IO ()

data IPAddrSource

Source from which the IP source address of the client is obtained.

Constructors

FromSocket
From the peer address of the HTTP connection.

FromHeader
From X-Real-IP: or X-Forwarded-For: in the HTTP header.