Network.Wai.Parse

Description

Some helpers for parsing data out of a raw WAI Request.

Documentation

parseHttpAccept :: ByteString -> [ByteString]

Parse the HTTP accept string to determine supported content types.

parseRequestBody :: BackEnd y -> Request -> ResourceT IO ([Param], [File y])

data RequestBodyType

Constructors

UrlEncoded

Multipart ByteString

getRequestBodyType :: Request -> Maybe RequestBodyType

sinkRequestBody :: BackEnd y -> RequestBodyType -> Sink ByteString (ResourceT IO) ([Param], [File y])

conduitRequestBody :: BackEnd y -> RequestBodyType -> Conduit ByteString (ResourceT IO) (Either Param (File y))

type BackEnd a

Arguments

= ByteString
parameter name

-> FileInfo ()

-> Sink ByteString (ResourceT IO) a

A file uploading backend. Takes the parameter name, file name, and content type, and returns a Sink for storing the contents.

lbsBackEnd :: Monad m => ignored1 -> ignored2 -> Sink ByteString m ByteString

Store uploaded files in memory

tempFileBackEnd :: MonadResource m => ignored1 -> ignored2 -> Sink ByteString m FilePath

Save uploaded files on disk as temporary files

tempFileBackEndOpts

Arguments

:: MonadResource m

=> IO FilePath
get temporary directory

-> String
filename pattern

-> ignored1

-> ignored2

-> Sink ByteString m FilePath

Same as tempFileSink, but use configurable temp folders and patterns.

type Param = (ByteString, ByteString)

Post parameter name and value.

type File y = (ByteString, FileInfo y)

Post parameter name and associated file information.

data FileInfo c

Information on an uploaded file.

Constructors

FileInfo

Fields

fileName :: ByteString

fileContentType :: ByteString

fileContent :: c