Network.Wai.EventSource.EventStream

Description

Internal module, usually you don't need to use it.

Documentation

data ServerEvent

Type representing a communication over an event stream. This can be an actual event, a comment, a modification to the retry timer, or a special "close" event indicating the server should close the connection.

Constructors

ServerEvent

Fields

eventName :: Maybe Builder

eventId :: Maybe Builder

eventData :: [Builder]

CommentEvent

Fields

eventComment :: Builder

RetryEvent

Fields

eventRetry :: Int

CloseEvent

eventToBuilder :: ServerEvent -> Maybe Builder

Converts a ServerEvent to its wire representation as specified by the text/event-stream content type.