Network.Wai.Middleware.Gzip

Description

Automatic gzip compression of responses.

Documentation

gzip :: GzipSettings -> Middleware

Use gzip to compress the body of the response.

Analyzes the "Accept-Encoding" header from the client to determine if gzip is supported.

Possible future enhancements:

Only compress if the response is above a certain size.

data GzipSettings

gzipFiles :: GzipSettings -> GzipFiles

data GzipFiles

Constructors

GzipIgnore

GzipCompress

GzipCacheFolder FilePath

gzipCheckMime :: GzipSettings -> ByteString -> Bool

def :: Default a => a

The default value for this type.

defaultCheckMime :: ByteString -> Bool