Network.Wai.Middleware.Gzip

Description

Automatic gzip compression of responses.

Documentation

gzip :: GzipSettings -> Middleware

Use gzip to compress the body of the response.

Analyzes the "Accept-Encoding" header from the client to determine if gzip is supported.

File responses will be compressed according to the GzipFiles setting.

data GzipSettings

gzipFiles :: GzipSettings -> GzipFiles

data GzipFiles

Gzip behavior for files.

Constructors

GzipIgnore
Do not compress file responses.

GzipCompress
Compress files. Note that this may counteract zero-copy response optimizations on some platforms.

GzipCacheFolder FilePath
Compress files, caching them in some directory.

GzipPreCompressed GzipFiles
If we use compression then try to use the filename with ".gz" appended to it, if the file is missing then try next action

Since: 3.0.17

gzipCheckMime :: GzipSettings -> ByteString -> Bool

def :: Default a => a

The default value for this type.

defaultCheckMime :: ByteString -> Bool

MIME types that will be compressed by default: text/*, application/json, application/javascript, application/ecmascript, image/x-icon.