Network.Wai.Middleware.MethodOverride

Documentation

methodOverride :: Middleware

Overriding of HTTP request method via _method query string parameter.

This middleware only applies when the initial request method is POST. Allows submitting of normal HTML forms, without worries of semantic mismatches with the HTTP spec.