Network.Wai.Middleware.MethodOverride

Documentation

methodOverride :: Middleware

Allows overriding of the HTTP request method via the _method query string parameter.

This middleware only applies when the initial request method is POST. This allow submitting of normal HTML forms, without worries of semantics mismatches in the HTTP spec.