Network.Wai.Middleware.RequestLogger

Basic stdout logging

logStdout :: Middleware

Production request logger middleware.

This uses the Apache logging format, and takes IP addresses for clients from the socket (see IPAddrSource for more information). It logs to stdout.

logStdoutDev :: Middleware

Development request logger middleware.

This uses the Detailed True logging format and logs to stdout.

Create more versions

mkRequestLogger :: RequestLoggerSettings -> IO Middleware

data RequestLoggerSettings

RequestLoggerSettings is an instance of Default. See Data.Default for more information.

outputFormat, autoFlush, and destination are record fields for the record type RequestLoggerSettings, so they can be used to modify settings values using record syntax.

defaultRequestLoggerSettings :: RequestLoggerSettings

outputFormat :: RequestLoggerSettings -> OutputFormat

Default value: Detailed True.

autoFlush :: RequestLoggerSettings -> Bool

Only applies when using the Handle constructor for destination.

Default value: True.

destination :: RequestLoggerSettings -> Destination

Default: Handle stdout.

data OutputFormat

The logging format.

Constructors

Apache IPAddrSource

ApacheWithSettings ApacheSettings
Since: 3.1.8

Detailed Bool
use colors?

DetailedWithSettings DetailedSettings
Since: 3.1.3

CustomOutputFormat OutputFormatter

CustomOutputFormatWithDetails OutputFormatterWithDetails

CustomOutputFormatWithDetailsAndHeaders OutputFormatterWithDetailsAndHeaders

data ApacheSettings

Settings for the ApacheWithSettings OutputFormat. This is purposely kept as an abstract data type so that new settings can be added without breaking backwards compatibility. In order to create an ApacheSettings value, use defaultApacheSettings and the various 'setApache' functions to modify individual fields. For example:

setApacheIPAddrSource FromHeader defaultApacheSettings

Since: 3.1.8

defaultApacheSettings :: ApacheSettings

setApacheIPAddrSource :: IPAddrSource -> ApacheSettings -> ApacheSettings

Where to take IP addresses for clients from. See IPAddrSource for more information.

Default value: FromSocket

Since: 3.1.8

setApacheRequestFilter :: (Request -> Response -> Bool) -> ApacheSettings -> ApacheSettings

Function that allows you to filter which requests are logged, based on the request and response

Default: log all requests

Since: 3.1.8

setApacheUserGetter :: (Request -> Maybe ByteString) -> ApacheSettings -> ApacheSettings

Function that allows you to get the current user from the request, which will then be added in the log.

Default: return no user

Since: 3.1.8

data DetailedSettings

Settings for the Detailed OutputFormat.

mModifyParams allows you to pass a function to hide confidential information (such as passwords) from the logs. If result is Nothing, then the parameter is hidden. For example: > myformat = Detailed True (Just hidePasswords) > where hidePasswords p@(k,v) = if k = "password" then (k, "***REDACTED***") else p

mFilterRequests allows you to filter which requests are logged, based on the request and response.

Since: 3.1.3

Constructors

DetailedSettings

Fields

useColors :: Bool

mModifyParams :: Maybe (Param -> Maybe Param)

mFilterRequests :: Maybe (Request -> Response -> Bool)

mPrelogRequests :: Bool

Since: 3.1.7

type OutputFormatter = ZonedDate -> Request -> Status -> Maybe Integer -> LogStr

type OutputFormatterWithDetails = ZonedDate -> Request -> Status -> Maybe Integer -> NominalDiffTime -> [ByteString] -> Builder -> LogStr

type OutputFormatterWithDetailsAndHeaders

Arguments

= ZonedDate
When the log message was generated

-> Request
The WAI request

-> Status
HTTP status code

-> Maybe Integer
Response size

-> NominalDiffTime
Duration of the request

-> [ByteString]
The request body

-> Builder
Raw response

-> [Header]
The response headers

-> LogStr

Same as OutputFormatterWithDetails but with response headers included

This is useful if you wish to include arbitrary application data in your logs, e.g., an authenticated user ID, which you would set in a response header in your application and retrieve in the log formatter.

Since: 3.0.27

data Destination

Constructors

Handle Handle

Logger LoggerSet

Callback Callback

type Callback = LogStr -> IO ()

data IPAddrSource

Source from which the IP source address of the client is obtained.

Constructors

FromSocket
From the peer address of the HTTP connection.

FromHeader
From X-Real-IP: or X-Forwarded-For: in the HTTP header.

FromFallback
From the peer address if header is not found.