Network.Wai.Middleware.RequestLogger.JSON

Documentation

formatAsJSON :: OutputFormatterWithDetails

formatAsJSONWithHeaders :: OutputFormatterWithDetailsAndHeaders

Same as formatAsJSON but with response headers included

This is useful for passing arbitrary data from your application out to the WAI layer for it to be logged, but you may need to be careful to subsequently redact any headers which may contain sensitive data.

Since: 3.0.27