Network.Wai.Middleware.StripHeaders

Documentation

stripHeader :: ByteString -> Response -> Response

stripHeaders :: [ByteString] -> Response -> Response

stripHeaderIf :: ByteString -> (Request -> Bool) -> Middleware

If the request satisifes the provided predicate, strip headers matching the provided header name.

Since 3.0.8

stripHeadersIf :: [ByteString] -> (Request -> Bool) -> Middleware

If the request satisifes the provided predicate, strip all headers whose header name is in the list of provided header names.

Since 3.0.8