Network.Wai.Test.Internal

Documentation

type Session = ReaderT Application (StateT ClientState IO)

type ClientCookies = Map ByteString SetCookie

Since 3.0.6

data ClientState

Constructors

ClientState

Fields

clientCookies :: ClientCookies

initState :: ClientState

Since 3.0.20.0

runSessionWith :: ClientState -> Session a -> Application -> IO (a, ClientState)

Like runSession, but if allows you to hand in cookies and get the updated cookies back. One use case for this is writing tests that address the application under test alternatingly through rest api and through db handle.

Since 3.0.20.0